requirements-dev.txt
wheel

argon2-cffi
passlib
authlib
psycopg2
pyjwt
redis

flask
flask-sqlalchemy
flask-admin
flask-socketio

pylint
pylint-flask
pylint-flask-sqlalchemy
flake8